mitra-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.8.3
groq==0.11.0
edge-tts==6.1.12
pydantic==2.9.0
python-multipart==0.0.12
